Final attempt to correct review's comments
