--- Concurrent trees/Coarse/Node.cs
@@ -1,16 +1,0 @@
-﻿#region
-
-#endregion
-
-namespace Concurrent_trees.Coarse
-{
-    public class Node
-    {
-        public int Key;
-        public Node Left;
-        public Node Parent;
-        public Node Right;
-
-        public Node(int key) => Key = key;
-    }
-}
